Nov 23/Discord Refrell Bot/requirements.txt
discord.py==2.3.2
google_api_python_client==2.96.0
oauth2client==4.1.3
